refactor: update favicons and image paths
