+++ src/components/Logo.astro
@@ -1,5 +1,5 @@
 ---
-import logo_transparent from '~/assets/images/logo_transparent.webp';
+import logo_transparent from '~/assets/images/other/logo_transparent.webp';
 import Image from '~/components/common/Image.astro';
 import { SITE } from 'astrowind:config';
 import GradientText from './ui/GradientText.astro';
