+++ src/components/ui/GradientText.astro
@@ -1,6 +1,6 @@
 ---
-import imgGold from '~/assets/images/gold2.webp';
-import imgGold2 from '~/assets/images/gold3.5.webp';
+import imgGold from '~/assets/images/gold_small.webp';
+import imgGold2 from '~/assets/images/gold.webp';
 import { twMerge } from 'tailwind-merge';
 
 const {
