+++ src/components/ui/ItemGrid.astro
@@ -1,5 +1,5 @@
 ---
-import imgGold from '~/assets/images/gold2.webp';
+import imgGold from '~/assets/images/gold_small.webp';
 import type { ItemGrid as Props } from '~/types';
 import { Icon } from 'astro-icon/components';
 import { twMerge } from 'tailwind-merge';
