+++ src/components/widgets/TestimonialCard.astro
@@ -1,5 +1,5 @@
 ---
-import imgGold2 from '~/assets/images/gold3.5.webp';
+import imgGold2 from '~/assets/images/gold.webp';
 import Image from '~/components/common/Image.astro';
 import { Icon } from 'astro-icon/components';
