+++ src/pages/index.astro
@@ -4,9 +4,9 @@ import ben from '~/assets/images/deacons/ben.webp';
 import ida from '~/assets/images/deacons/ida.webp';
 import cj from '~/assets/images/elders/cj.webp';
 import mahi from '~/assets/images/elders/mahi.webp';
-import prayerConference from '~/assets/images/Kuopio Prayer conference_ April 2025_finnish.webp';
-import logo from '~/assets/images/logo.webp';
-import imgPrayerChurchKuopio from '~/assets/images/prayer_church_kuopio2.webp';
+import prayerConference from '~/assets/images/other/Kuopio Prayer conference_ April 2025_finnish.webp';
+import logo from '~/assets/images/other/logo.webp';
+import imgPrayerChurchKuopio from '~/assets/images/other/prayer_church_kuopio2.webp';
 import iina from '~/assets/images/testimonials/iina.webp';
 import maria from '~/assets/images/testimonials/maria.webp';
 import GradientText from '~/components/ui/GradientText.astro';
